refactor(runtime): keep image module under cap
